separating search and details into 2 components
