+++ src/components/SideBar/PlaceDetailsView.js
@@ -1,0 +1,10 @@
+const PlaceDetailsView = ({ selectedPlace, onGoBack }) => {
+    return (
+        <div>
+            <h1>{selectedPlace.name}</h1>
+            <button onClick={onGoBack}>go back</button>
+        </div>
+    );
+};
+
+export default PlaceDetailsView;
